feat: isPurchased filter middleware
