+++ src/VirtoCommerce.XOrder.Core/VirtoCommerce.XOrder.Core.csproj
@@ -9,9 +9,9 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
-    <PackageReference Include="VirtoCommerce.OrdersModule.Core" Version="3.844.2" />
+    <PackageReference Include="VirtoCommerce.OrdersModule.Core" Version="3.844.0-alpha.1321-vcst-3137" />
     <PackageReference Include="VirtoCommerce.Xapi.Core" Version="3.906.0" />
     <PackageReference Include="VirtoCommerce.XCart.Core" Version="3.906.0" />
-    <PackageReference Include="VirtoCommerce.XCatalog.Core" Version="3.911.3" />
+    <PackageReference Include="VirtoCommerce.XCatalog.Core" Version="3.911.0-alpha.89-vcst-3137-purchased-before" />
   </ItemGroup>
 </Project>
