+++ src/VirtoCommerce.XOrder.Data/VirtoCommerce.XOrder.Data.csproj
@@ -11,7 +11,7 @@
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
     <PackageReference Include="VirtoCommerce.FileExperienceApi.Core" Version="3.904.0" />
     <PackageReference Include="VirtoCommerce.Platform.Security" Version="3.877.0" />
-    <PackageReference Include="VirtoCommerce.XCatalog.Data" Version="3.911.3" />
+    <PackageReference Include="VirtoCommerce.XCatalog.Data" Version="3.911.0-alpha.89-vcst-3137-purchased-before" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\VirtoCommerce.XOrder.Core\VirtoCommerce.XOrder.Core.csproj" />
